feat(show-data): create new branch show-data from storage-matrix

+++ src-tauri/src/events.rs
@@ -13,7 +13,7 @@ pub struct AFKEvent {
 
 #[derive(Debug)]
 pub struct WindowInformation {
-    pub time: u128,
+    pub time: u64,
     pub title: Option<String>,
     pub class: Option<Vec<String>>,
     pub exec_path: Option<String>,
